Added the concept of "block sounds", which are defined in blockSounds/*.blocksounds.
The idea here is that large groups of blocks will have the same soundscape. Instead of repeating the sound definitions in each block (or template), a soundscape is defined once and then referenced via ID from the block definition. Sounds from the given lists are randomly played when that event occurs.

For footsteps, the block's footstep sounds are prioritized over the creature's footsteps at this point. This probably needs more work when other types of entities are created.