Clear the interactionTarget field when an interaction dialog gets closed.
A module developer that usess the InteractionScreen component should not
need to know about the need to update the interactionTarget field
of the character.

This patch makes will make automatically opened interaction screens,
also automatically update the interactionTarget field when they get
closed.

For that purpose a new event is introduced called ScreenLayerClosedEvent
which gets sent to the client component when a screen layer gets closed.