Big unfinished change to save transactions:
The game takes now a snapshot on the main thread and writes it in a
background thread.

Performance isn't great yet: All chunks are encoded in the main thread.

The purgeWorld command got replaced by a save command, as the former
need to be reimplemented from scratch.

The chunk unloading has been disabled for now,
until it is ported to the new system.

Disconnecting players don't get handled properly yet, as their
state saving needs to be ported to the new system too.