Prevent crashes from corrupting the save game.
Changes to the save game get now placed in a subdirectory.
When there is an interruption before new versions of the files could
be generated in the directory then the directory will be deleted at
startup.
When the changes are complete, the directory will be renamed.
The name then indicates that the files should be copied over
the existing files. If it gets interrupted it can be resumed
at the next the world gets loaded.