Don't use the float suffix f in post_frag.glsl.
Reasoning:

This will fix error messages like the following generated by a buggy
graphic driver:

"Only GLSL version > 110 allows postfix "F" or "f" for float
WARNING: 0:475: Only GLSL version > 110 allows postfix "F" or "f" for float"

In GLSL shader version 1.10 it is not allowed to use the suffix,
but we explictly state via #version that version 1.20 should be used.
So it is a bug if a graphic driver complains about it.

GLSLShader#createShaderBuilder