Fix block plaement near player.
Without this patch it is not possible to place blocks in certain
situations:
1. You couldn't place a block at character height if you walked towards
another path blocking block that is below or above the placement position.
e.g. place a block infront of you, walk towards it till limit.
Try placing a block on top of it.

2. Blocks couldn't be placed in a neighbour block that is 1 lower than
the current character position even when the character wasn't lowered
already from the edge (and capsule collision shape).

If in doubt the block placement will now be allowed which can lead in
certain corner situations to blocks that collide with the character.
e.g. when you walk very far on the edge and then plae a block below you.
You will then sink in the block till you hit the next block below it so
it is no big deal.