Send chunks to client before light calculation is done.
Reasoning:

The light calculation gets only performed once the adjacent chunks got
loaded. So if wait for the light calculation before we mark a chunk as
relevant for a client then we would transfer less chunks to the client
then the relevance region is large. the client would then again perform
the light calculation too based on that reduced chunk count and would
reduce the view distance again. That is why it makes sense to detect
chunks as relevant even when no light calculation has been performed yet.